+++ pyproject.toml
@@ -14,6 +14,10 @@ grobid-tei-xml = "^0.1.3"
 pyinstaller = "^5.11.0"
 ipython = "^8.13.2"
 
+
+[tool.coverage.run]
+omit = "*/tests/*"
+
 [build-system]
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
