+++ python/sidecar/functions.py
@@ -1,0 +1,2 @@
+def get_word_count(text: str) -> int:
+    return len(text.strip().split(" "))
